feat: model refactoring
