+++ src/main/java/org/broadinstitute/consent/http/models/support/CustomRequestField.java
@@ -1,47 +1,10 @@
 package org.broadinstitute.consent.http.models.support;
 
-import java.util.Objects;
+/**
+ * Represents custom fields in a Zendesk support ticket <a
+ * href="https://developer.zendesk.com/api-reference/ticketing/tickets/ticket-requests/">Requests
+ * API Reference</a>
+ */
+public record CustomRequestField(long id, String value) {
 
-public class CustomRequestField {
-
-  private long id;
-  private String value;
-
-  public CustomRequestField(long id, String value) {
-    this.id = id;
-    this.value = value;
-  }
-
-  public long getId() {
-    return id;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
-  public String getValue() {
-    return value;
-  }
-
-  public void setValue(String value) {
-    this.value = value;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    CustomRequestField that = (CustomRequestField) o;
-    return Objects.equals(id, that.id) && Objects.equals(value, that.value);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, value);
-  }
 }
