+++ src/main/java/org/broadinstitute/consent/http/models/support/SupportRequester.java
@@ -1,32 +1,10 @@
 package org.broadinstitute.consent.http.models.support;
 
 /**
- * Represents user creating a ticket to request support via Zendesk
- * https://broadinstitute.zendesk.com/api/v2/requests.json
+ * Represents requester creating a ticket to request support via Zendesk <a
+ * href="https://developer.zendesk.com/api-reference/ticketing/tickets/ticket-requests/">Requests
+ * API Reference</a>
  */
-public class SupportRequester {
+public record SupportRequester(String name, String email) {
 
-  private String name;
-  private String email;
-
-  public SupportRequester(String name, String email) {
-    this.name = name;
-    this.email = email;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
-  }
 }
