+++ src/test/java/org/broadinstitute/consent/http/AbstractTest.java
@@ -1,0 +1,24 @@
+package org.broadinstitute.consent.http;
+
+import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.RandomUtils;
+
+public abstract class AbstractTest {
+
+  public String randomAlphabetic(int length) {
+    return RandomStringUtils.secureStrong().nextAlphabetic(length);
+  }
+
+  public String randomAlphanumeric(int length) {
+    return RandomStringUtils.secureStrong().nextAlphanumeric(length);
+  }
+
+  public int randomInt(int startInclusive, int endExclusive) {
+    return RandomUtils.secureStrong().randomInt(startInclusive, endExclusive);
+  }
+
+  public long randomLong() {
+    return RandomUtils.secureStrong().randomLong();
+  }
+
+}
